Update WELSPECS.fodt
When the general well specification data is defined by the WELSPECS keyword in the SCHEDULE section the well can now be assigned directly to the FIELD group in item 2 GRPNAME (#3485 and #4608). Previously there was a restriction preventing wells from being parented directly to FIELD. Although wells are now allowed to be parented directly to FIELD, this is discouraged and a warning message will be issued. Mixing wells and groups as children of a single group is still forbidden.