Update GCONPROD for 2023-10
(1) Added support for GCONPROD item 7 ACTION equal to NONE in the SCHEDULE section, which specifies that no action is to be taken if the oil, water, gas or liquid rate constraints are violated. Previously only the RATE option was supported by the simulator.
(2) Added partial support for setting items 11, 12 and 13 of GCONPROD in the SCHEDULE section: ACTWAT, ACTGAS and ACTLIQ respectively define the action to be taken if the water rate (WRAT), gas rate (GRAT) and liquid rate (LRAT) constraints defined by GCONPROD are violated. Supported options are now NONE or RATE. Options CON, +CON, WELL, and PLUG are not currently supported (#4748).