Update WPAVE for 2023-10
Added partial support for WPAVE and WWPAVE defining the method and parameters for calculating a well’s block average pressures for either all wells or specific wells (#4695, #4694 and #4693). Calculation of block averages pressures is currently only supported for OPEN completions (not for ALL completions) as specified in WPAVE item 4 and WWPAVE item 5.