requested changes
If possible please make sure the fonts are embedded (on File->Properties->Fonts please check the first two boxes). Otherwise the page breaks make no sense on my machine when looking at the keyword file. It should be OK when looking at the full file already now. (@hakonhagland can we change the script to do this?)
In the reference there is a trailing 1 on "Sandve1". Should be Sandve.
In the example, fourth listing the line after RUNSPECT does not seem to be properly indented (uses spaces instead)