Added a raw bash script to review differences made to the text.
scripts/text-diff.sh is totally unpolished and might fail in
various occasions. Still, it rather handy to detect and review
differences made to the text in a subdocument.

Arguments are the PR number and the base name of the changed file.
It will checkout main, convert the file to txt and store it as
basename_old.txt in /tmp. Then it will checkout the PR, convert
the updated document to txt in file /tmp/basename.txt. Execute diff
on those two files and start liberoffice with the main file.