Escape XML quotes and apostrophes
xml.sax.saxutils.escape() escapes only the three characters "&",
"<", ">". But LibreOffice also escapes the characters '"', "'" so
we should escape them as well in order to not get large PR diffs
each time a script is modifying the XML files. Or each time
LibreOffice is modifying an XML file that was previously written
by a Python script.