Create script to extract keyword URIs
Python script to extract a mapping from keyword names to XML URI tag
names. The mapping can be saved to a file and later used to make
keywords clickable. That is, to be able to click on a keyword referenced
in a paragraph of text, and then jump to the page where that keyword
is defined.